// frontend/src/lib/api/generated/models/ListingFeedbacksItem.ts
/**
 * Roominar
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { UserFeedbackRating } from '../models/UserFeedbackRating';
import { HttpFile } from '../http/http';

export class ListingFeedbacksItem {
    'id': number;
    'positiveFeedback'?: string | null;
    'negativeFeedback'?: string | null;
    'createdAt': Date;
    'updatedAt': Date;
    'isAnonymous'?: boolean;
    'userId'?: number | null;
    'userName'?: string | null;
    'userAvatar'?: string | null;
    'deletedAt'?: Date | null;
    'ratings'?: Array<UserFeedbackRating>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "positiveFeedback",
            "baseName": "positive_feedback",
            "type": "string",
            "format": ""
        },
        {
            "name": "negativeFeedback",
            "baseName": "negative_feedback",
            "type": "string",
            "format": ""
        },
        {
            "name": "createdAt",
            "baseName": "created_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "updatedAt",
            "baseName": "updated_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "isAnonymous",
            "baseName": "is_anonymous",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "userId",
            "baseName": "user_id",
            "type": "number",
            "format": ""
        },
        {
            "name": "userName",
            "baseName": "user_name",
            "type": "string",
            "format": ""
        },
        {
            "name": "userAvatar",
            "baseName": "user_avatar",
            "type": "string",
            "format": ""
        },
        {
            "name": "deletedAt",
            "baseName": "deleted_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "ratings",
            "baseName": "ratings",
            "type": "Array<UserFeedbackRating>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ListingFeedbacksItem.attributeTypeMap;
    }

    public constructor() {
    }
}